+++ apps/drupal/webpack.drupal.prod.js
@@ -9,7 +9,7 @@ const merge = require('webpack-merge');
 const webpack = require('webpack');
 
 // Custom Imports
-const shared = require('../../webpack.shared.prod.js');
+const shared = require('../../webpack.particle.prod.js');
 
 const drupal = {
   mode: 'production',
